+++ src/resources/scss/_variables.scss
@@ -1,0 +1,1 @@
+$red-color: red;
