+++ src/resources/scss/app.scss
@@ -1,0 +1,9 @@
+@import 'variables';
+
+h1 {
+    color: $red-color;
+}
+
+h2{
+    color: green;
+}
