+++ src/resources/views/home.hbs
@@ -1,0 +1,1 @@
+<h1>Hello Moi nguoi</h1>
