+++ src/resources/views/news.hbs
@@ -1,0 +1,1 @@
+<h1>News page</h1>
